+++ force_bdss/cli/force_bdss.py
@@ -1,13 +1,12 @@
 import click
 from envisage.core_plugin import CorePlugin
-import logging
 
 from force_bdss.bdss_application import BDSSApplication
 from force_bdss.core_mco_driver import CoreMCODriver
-from force_bdss.multi_criteria_optimizers_plugin import \
-    MultiCriteriaOptimizersPlugin
-from force_bdss.key_performance_calculators_plugin import \
+from force_bdss.kpi.key_performance_calculators_plugin import \
     KeyPerformanceCalculatorsPlugin
+from force_bdss.mco.multi_criteria_optimizers_plugin import \
+    MultiCriteriaOptimizersPlugin
 
 
 @click.command()
