+++ force_bdss/kpi/basic_key_performance_calculator.py
@@ -1,6 +1,6 @@
 from traits.api import provides, HasStrictTraits
 
-from force_bdss.i_key_performance_calculator import IKeyPerformanceCalculator
+from force_bdss.kpi.i_key_performance_calculator import IKeyPerformanceCalculator
 
 
 @provides(IKeyPerformanceCalculator)
